Remove unused entity processors (`ClosureEntityProcessor`, `PassThroughEntityProcessor`), `NotImplementedException` class, update exceptions in `UsersRepository`, and refactor `UsersServiceInterface` namespace for better alignment.
