--- src/Exception/NotImplementedException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace ByJG\Authenticate\Exception;
-
-use Exception;
-
-class NotImplementedException extends Exception
-{
-    //put your code here
-}
